% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_uncertainty_in_obs.R
\name{add_uncertainty_in_obs}
\alias{add_uncertainty_in_obs}
\title{Add uncertainty to observations}
\usage{
add_uncertainty_in_obs(fit, perc_low = 0.025, perc_up = 0.975)
}
\arguments{
\item{fit}{model_fit object.}

\item{perc_low}{lower percentile for uncertainty.}

\item{perc_up}{upper percentile for uncertainty.}
}
\description{
This function adds uncertainty to observations using the model fit object.
}
\keyword{internal}
