% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outliers.R
\name{assign_outliers}
\alias{assign_outliers}
\title{Assign what observations are given an outlier error in fit
to do: rename as assign_outlier_errors
consider case_when or so in this function!
test that it works with and without outlier_record_ids}
\usage{
assign_outliers(dat, outlier_record_ids = NULL)
}
\arguments{
\item{dat}{data, columns possible_outlier and possible_outlier_userinput are used if included,
if not, we use an algorithm that depends on obs year, data series type, and iso}

\item{outlier_record_ids}{if provided, the obs with these record_ids are assigned to be possibly outlying.}
}
\value{
dat with nooutlier column added
}
\description{
Assign what observations are given an outlier error in fit
to do: rename as assign_outlier_errors
consider case_when or so in this function!
test that it works with and without outlier_record_ids
}
