% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_data}
\alias{explore_data}
\title{Explore Coverage Indicators Over Time}
\usage{
explore_data(
  data,
  indicator_col,
  indicator_se_col = NULL,
  data_types_col = NULL,
  group_col,
  region_col = NULL,
  indicator_name = NULL,
  routine_data = NULL
)
}
\arguments{
\item{data}{A data frame containing the coverage indicators and year variable.}

\item{indicator_col}{A string specifying the column name for the coverage indicator.}

\item{indicator_se_col}{An optional string specifying the column name for the standard error of the indicator.}

\item{data_types_col}{An optional string specifying the column name for data types in the data frame.}

\item{group_col}{A string specifying the column name for first-level grouping (e.g., country).}

\item{region_col}{An optional string specifying the column name for second-level grouping (e.g., region).}

\item{indicator_name}{An optional string to use as the title for plots instead of the default "Indicator".}

\item{routine_data}{An optional data frame containing routine data.}
}
\value{
A series of plots for exploratory data analysis.
}
\description{
This function performs exploratory data analysis on a dataset containing coverage indicators over time.
}
