% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit the transition model to data.}
\usage{
fit_model(
  survey_df,
  national_dat_df = NULL,
  routine_df = NULL,
  mean_log_sigma = NULL,
  popweights = NULL,
  y = "invprobit_indicator",
  se = "se_invprobit_indicator",
  year = "year",
  source = "data_series_type",
  area = "iso",
  iso_select = NULL,
  routine_data = NULL,
  start_year = 2000,
  end_year = 2030,
  runstep,
  global_fit = NULL,
  t_star = 2010,
  spline_degree = 2,
  num_knots = 8,
  Betas_upper_bound = 0.5,
  Betas_lower_bound = 0.01,
  Ptilde_low = 0,
  add_dataoutliers = TRUE,
  extra_stan_data = list(),
  hierarchical_level = c("intercept", "cluster", "subcluster", "iso"),
  hierarchical_splines = c("intercept", "cluster", "subcluster", "iso"),
  hierarchical_asymptote = c("intercept", "iso"),
  add_subnational_hierarchy = "admin1",
  use_globalsubnat_fromnat = FALSE,
  held_out = FALSE,
  validation_cutoff_year = NULL,
  validation_run = FALSE,
  save_post_summ = FALSE,
  generate_quantities = TRUE,
  get_posteriors = TRUE,
  create_runname_and_outputdir = TRUE,
  runnumber = 1,
  rungroup = NULL,
  runname = NULL,
  chains = 4,
  iter_sampling = 200,
  iter_warmup = 150,
  add_sample = TRUE,
  compile_model = TRUE,
  force_recompile = FALSE,
  seed = 1234,
  refresh = 10,
  adapt_delta = 0.9,
  max_treedepth = 14,
  variational = FALSE,
  nthreads_variational = 8,
  add_inits = TRUE
)
}
\arguments{
\item{survey_df}{tibble with survey data}

\item{routine_df}{tibble with routine data}

\item{y}{column name of outcome.}

\item{se}{column name of outcome standard error.}

\item{year}{column name of outcome year.}

\item{source}{column name of data source.}

\item{area}{column name of the area of each observation

for subnational:}

\item{iso_select}{ISO code to use for local national run}

\item{start_year}{start year of estimates.}

\item{end_year}{end year of estimates.

Settings for global model fit 1a}

\item{runstep}{type of run, currently one of "step1a", "step1ab", "step1b", "local_national" (see Details).}

\item{global_fit}{optional object of class \code{"fpemplus"}, used to obtain fixed
values to use for some parameters in the current fit (see Details).}

\item{t_star}{reference year used in model.}

\item{spline_degree}{spline degree. Degree 2 or 3 is supported.}

\item{num_knots}{number of spline knots.}

\item{Betas_upper_bound}{upper bound for the splines parameters}

\item{Betas_lower_bound}{lower bound for the splines parameters}

\item{Ptilde_low}{lower bound for the asymptote Ptilde}

\item{add_dataoutliers}{boolean indicator of whether to include data outliers in 1b}

\item{extra_stan_data}{list of additional data to pass to Stan model}

\item{hierarchical_level}{vector specifying hierarchical structure for the
level in reference year (see Details).}

\item{hierarchical_splines}{vector specifying hierarchical structure for
spline coefficients (see Details).}

\item{hierarchical_asymptote}{vector specifying hierarchical structure for
asymptote (see Details).}

\item{held_out}{binary vector indicating which observations are held out. Set to FALSE to hold out no observations.}

\item{validation_cutoff_year}{year to use for out-of-sample validation, overwrites held_out (to confirm it does)}

\item{validation_run}{boolean indicator of whether it's a validation model run or not}

\item{generate_quantities}{binary vector indicating whether to simulate data from the fitted model

Setting for where to save things}

\item{get_posteriors}{boolean indicator of whether to return posterior samples}

\item{create_runname_and_outputdir}{boolean indicator of whether to create a runname and output directory}

\item{runnumber}{number to add to runname}

\item{rungroup}{group to add to runname}

\item{runname}{name to use for run

Settings for sampling}

\item{chains}{number of chains to run}

\item{iter_sampling}{number of posterior samples to draw}

\item{iter_warmup}{number of warmup iterations}

\item{add_sample}{boolean indicator of whether to return samples}

\item{compile_model}{boolean indicator of whether to compile the Stan model}

\item{force_recompile}{boolean indicator of whether to force recompilation of the Stan model}

\item{seed}{random seed}

\item{refresh}{number of iterations between progress updates}

\item{adapt_delta}{target acceptance rate for the No-U-Turn Sampler}

\item{max_treedepth}{maximum tree depth for the No-U-Turn Sampler}

\item{variational}{boolean indicator of whether to use variational inference (not yet tested)}

\item{nthreads_variational}{number of threads to use for variational inference}

\item{add_inits}{boolean indicator of whether to add initial values to the Stan model}

\item{population_data}{a data frame with yearly population counts for
subnational regions. It should have columns matching the names specified
for \code{year} and \code{area}. This data frame is only required if the
primary \code{data} set contains a mix of observations at national and
subnational levels.}
}
\value{
fpemplus object.
}
\description{
Fit the transition model to data.
}
\details{
The \code{fit_model} function fits the transition model to data. The model is
fitted using Stan, and the function returns an object of class \code{"fpemplus"}.
The argument \code{runstep} determines the type of run to perform. The
following run steps are supported:
\itemize{
\item "step1a": Fit the model without the smoothing term, to estimate longer term trends.
\item "step1ab": Fit the model with smoothing terms, estimating all parameters (none fixed).
Unlike step1b, this does not require a prior fit and estimates everything in one pass.
\item "step1b": Fit the model with smoothing terms, using a fit from step1a,
to estimate all smoothing and data model parameters.
\item "local_national": Fit the model to data from a single country, using a 1b fit.
This is also explained in the documentation folder.
}

Details on hierarchical set ups used
Several area-specific parameters of the fpemplus model have hierarchical priors
assigned to them so that information can be shared between areas.
The package allows the structure of the hierarchical prior to be configured by the user
through the \code{hierarchical_asymptote}, \code{hierarchical_level}, and \code{hierarchical_splines} arguments.
These arguments expect a character vector that specifies a nesting hierarchical structure.
Each element of the vector must be either "intercept" or a column name in the dataset, where
"intercept" will add a global intercept for the parameter.
The vector must be in descending order in terms of the hierarchy: that is, it starts with
"intercept" and proceeds down the hierarchy.

For example, suppose we are fitting country-level data, where the dataset has columns
"name_country", "name_sub_region", and "name_region" containing the name of the country,
sub-region, and region that each observation belongs to. To specify that the spline coefficients
should be fitted with a hierarchical model in which countries are nested within sub-regions within regions within world,
we would use the argument
\code{hierarchical_splines = c("intercept", "name_region", "name_sub_region", "name_country")}.

Optionally, model parameters can be fixed to values from a previous model fit
provided via the \code{global_fit} argument. In a typical use case, the
\code{global_fit} will have been fit to data from many geographic units
(e.g., all countries), while the current fit uses data from a smaller number
of locations.
Note: remainder of details is currently determined by the runstep
To use a global fit to fix parameter values, a number of settings must be the
same in the global fit and the current call to \code{fpemplus}:
\itemize{
\item For any of the hierarchical settings, e.g.
\code{hierarchical_asymptote}, all of the hierarchical levels
used for that parameter in the global fit must also be used in the current
fit. For instance, if the \code{global_fit} used
\code{hierarchical_asymptote = c("intercept", "name_region", "name_country")},
then in the current fit it is valid to use
\code{hierarchical_asymptote = c("intercept", "name_region", "name_country")} again or
\code{hierarchical_asymptote = c("intercept", "name_region", "name_country", "name_subnational")},
but it is not valid to use
\code{hierarchical_asymptote = c("intercept", "name_region", "name_subnational")}.
\item All hierarchical levels for terms and sigmas to fix in the current fit are
contained in the levels in the hierarchy used for the corresponding
parameters in the global fit. For example, if
\code{hierarchical_asymptote_sigmas_fixed = c("intercept", "name_region", "name_country")}
then the global fit must have included at least "intercept", "name_region",
and "name_country" for \code{hierarchical_asymptote}.
\item Any hierarchical levels to fix in the current fit are at the highest levels
of the hierarchical structure. For example, if
\code{hierarchical_asymptote = c("intercept", "name_region", "name_country")},
then it is valid to use
\code{hierarchical_asymptote_sigmas_fixed = c("intercept", "name_region")},
but it is not valid to use
\code{hierarchical_asymptote_sigmas_fixed = c("intercept", "name_country")}.
\item It is not valid to fix terms at a given hierarchy level without also fixing
the sigma estimate at that hierarchy level. For example, we cannot specify
\code{hierarchical_asymptote_sigmas_fixed = c("intercept")} and
\code{hierarchical_asymptote_terms_fixed = c("intercept", "name_region")}
\item It is only valid to set \code{fix_smoothing = TRUE} if also \code{smoothing = TRUE}.
\item All settings for the arguments \code{model}, \code{t_star}, \code{smoothing}, \code{tau_prior},
and \code{rho_prior} must be the same in the \code{global_fit} and the current fit.
\item If \code{hierarchical_splines_sigmas_fixed} or
\code{hierarchical_splines_terms_fixed} include any hierarchical levels (i.e.,
if either is different from the empty vector \code{c()}), all settings for
\code{num_knots} and \code{spline_degree} must be the same in the \code{global_fit} and
the current fit.
\item All geographic units that appear in the \code{data} for the current fit at
hierarchical levels for which any parameter is fixed must have also been
included in the \code{data} used for the \code{global_fit}.
\item If \code{fix_nonse = TRUE}, all data \code{source}s that appear in the \code{data} for the
current fit must also have been included in the \code{data} used for the
\code{global_fit}.
}
}
