% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_routine.R
\docType{data}
\name{fit_routine}
\alias{fit_routine}
\title{Fit routine}
\format{
brm fitted object
}
\usage{
fit_routine
}
\description{
Fitted brm object with information on routine sigmas.
}
\keyword{datasets}
