% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standata_routine.R
\name{get_logsigma_mean}
\alias{get_logsigma_mean}
\title{Routine data mean uncertainty}
\usage{
get_logsigma_mean(fit_routineglobal, indicator, worst_combi)
}
\arguments{
\item{fit_routineglobal}{brms fit for error variance routine data}

\item{indicator}{name indicator}

\item{worst_combi}{vector with value of data quality indicator,
worst of start and end year for annual change}
}
\value{
vector with mean log sigma
}
\description{
Routine data mean uncertainty
}
\keyword{internal}
