% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_helpers.R
\name{get_se_invprobitprop}
\alias{get_se_invprobitprop}
\title{Get Standard Error on invProbit Scale from Proportion}
\usage{
get_se_invprobitprop(prop, se_prop)
}
\arguments{
\item{prop}{Proportion value (between 0 and 1)}

\item{se_prop}{Standard error on proportion scale}
}
\value{
Standard error on probit scale
}
\description{
Transforms standard error from proportion scale to invprobit scale using the delta method.
Uses the derivative of the inverse probit function: SE_probit = SE_prop / dnorm(qnorm(prop))
}
