% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_helpers.R
\name{get_se_logitprop}
\alias{get_se_logitprop}
\title{Get Standard Error on Logit Scale}
\usage{
get_se_logitprop(prop, se_prop)
}
\arguments{
\item{prop}{Proportion value (between 0 and 1)}

\item{se_prop}{Standard error on proportion scale}
}
\value{
Standard error on logit scale
}
\description{
Transforms standard error from proportion scale to logit scale using the delta method.
}
