% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_helpers.R
\name{get_se_probitofinvprobitprop}
\alias{get_se_probitofinvprobitprop}
\title{Get Standard Error on Proportion Scale from invProbit proportion}
\usage{
get_se_probitofinvprobitprop(invprobitprop, se_invprobitprop)
}
\arguments{
\item{invprobitprop}{Value on invprobit scale (i.e., qnorm of proportion)}

\item{se_invprobitprop}{Standard error on invprobit scale}
}
\value{
Standard error on proportion scale
}
\description{
Transforms standard error from invprobit scale back to proportion scale using the delta method.
Uses the derivative of the probit function: SE_prop = SE_invprobit * dnorm(invprobit_value)
}
