% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standata_routine.R
\name{get_standata_routine}
\alias{get_standata_routine}
\title{get_standata_routine}
\usage{
get_standata_routine(routine_data, fit_routine_obj, time_index, geo_unit_index)
}
\arguments{
\item{routine_data}{tibble with  iso, year, indicator_name, routine_value, countdownmean
routine_value is the value of the routine data (eg coverage) and
countdownmean is the data quality indicator for that iso-year-indicator combi
needs to be filtered to iso of interest}

\item{fit_routine_obj}{brm-fit object}

\item{time_index}{from the model fit}

\item{geo_unit_index}{from the model fit}
}
\value{
list with dat_routine (for plotting) and routine_list (to pass to stan)
}
\description{
get_standata_routine
}
\keyword{internal}
