% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_fit_objects.R
\docType{data}
\name{globalfit_global_subnational_cci}
\alias{globalfit_global_subnational_cci}
\title{Global subnational fit object globalfit_global_subnational_cci}
\format{
\subsection{\code{globalfit_global_subnational_cci}}{

list with global fit objects including
\describe{
\item{post_summ}{data frame with posterior summaries for relevant parameters}
\item{geo_unit_index}{data frame with geo_unit info}
\item{source_index}{data frame with sources of survey data with indices used in the model}
\item{time_index}{data frame with TIME info}
\item{other settings}{spline settings, t_star, etc}
\item{data_from2010only}{logical indicating to fit to data from 2010 onwards only (TRUE) or with all available data (FALSE)}
}
}
}
\source{
global fit 2026
}
\usage{
globalfit_global_subnational_cci
}
\description{
Global subnational fit object globalfit_global_subnational_cci
}
\keyword{internal}
