% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_wrappers.R
\name{has_cmdstanr}
\alias{has_cmdstanr}
\title{Check if cmdstanr is available}
\usage{
has_cmdstanr()
}
\value{
Logical indicating if cmdstanr package is installed
}
\description{
Check if cmdstanr is available
}
\keyword{internal}
