% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_outliers.R
\name{identify_outliers_in_global_fit}
\alias{identify_outliers_in_global_fit}
\title{identify_outliers_in_global_fit}
\usage{
identify_outliers_in_global_fit(fit)
}
\arguments{
\item{fit}{Global fit object with posterior etas and data set that has record_id_fixed}
}
\value{
vector with record_id_fixed of outliers
}
\description{
identify_outliers_in_global_fit
}
\keyword{internal}
