% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inits.R
\name{init_fun}
\alias{init_fun}
\title{Inits function}
\usage{
init_fun(chain_id, stan_data)
}
\arguments{
\item{chain_id}{Chain ID}

\item{stan_data}{Stan data list, used to determine which parameters to initialize and their dimensions}
}
\value{
List of initial values for Stan parameters, with dimensions matching the model specification
}
\description{
Inits function
}
\keyword{internal}
