% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_helpers.R
\name{inv_probit}
\alias{inv_probit}
\title{Inverse Probit Transform (Proportion to InvProbit)}
\usage{
inv_probit(x)
}
\arguments{
\item{x}{Value on proportion scale (between 0 and 1)}
}
\value{
Value on inv-probit scale
}
\description{
Transforms values from proportion scale to inv-probit scale using the inverse standard normal CDF.
}
