% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_fits.R
\name{parse_runname}
\alias{parse_runname}
\title{Parse a run name into its components}
\usage{
parse_runname(runname)
}
\arguments{
\item{runname}{Character string of format {indicator}\emph{{runstep}}{folder_suffix}}
}
\value{
List with indicator, runstep, and folder_suffix
}
\description{
Parse a run name into its components
}
\keyword{internal}
