% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates_local.R
\name{plot_estimates_local}
\alias{plot_estimates_local}
\title{Plot local model fits.}
\usage{
plot_estimates_local(
  estimates,
  filtered_data,
  estimates2 = NULL,
  estimates3 = NULL,
  estimates4 = NULL,
  modelnames,
  indicator_name,
  add_estimates = TRUE,
  all_survey_types = NULL,
  cols_sourcetypes = c(DHS = "red", DHS0 = "red", MICS = "blue", PMA = "darkgreen", Other
    = "orange", `National survey` = "purple", NSS = "deepskyblue2", `Routine data` =
    "black")
)
}
\arguments{
\item{estimates}{A data frame estimates from the model fit.}

\item{filtered_data}{A data frame containing data from the model fit.}

\item{estimates2}{Optional estimates from a second model.}

\item{estimates3}{Optional estimates from a third model.}

\item{estimates4}{Optional estimates from a fourth model.}

\item{modelnames}{A vector of model names used for plotting. Default is \code{c("model1", "model2", "model3", "model4")}.}

\item{indicator_name}{Name of indicator used for plotting.}

\item{add_estimates}{Boolean to show model estimates/ribbons.}

\item{all_survey_types}{Optional character vector of all survey types to include in legend (for consistent legends across multiple plots).}

\item{cols_sourcetypes}{A named vector of colors for the different data series types (e.g., DHS, MICS, etc.).}
}
\description{
Plot local model fits.
}
