% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates_local.R
\name{plot_estimates_local_all}
\alias{plot_estimates_local_all}
\title{Plot model fits, with options to add routine data or additional fits for comparison.}
\usage{
plot_estimates_local_all(
  results,
  dat_routine = NULL,
  indicator_name = NULL,
  results2 = NULL,
  results3 = NULL,
  results4 = NULL,
  modelnames = c("model1", "model2", "model3", "model4"),
  iso_codes = NULL,
  save_plots = FALSE,
  output_folder = NULL,
  add_caption = FALSE,
  add_estimates = TRUE,
  use_for_facetting = FALSE
)
}
\arguments{
\item{results}{Model fit.}

\item{dat_routine}{Optional routine data to add to model estimates for comparison.}

\item{indicator_name}{Custom title for the y label of the plot. If \code{NULL}, name will be pulled from model fit.}

\item{results2}{Optional second model fit.}

\item{results3}{Optional third model fit.}

\item{results4}{Optional fourth model fit.}

\item{modelnames}{Optional names of the models to display in the plot legend.}

\item{iso_codes}{Optional character vector of ISO codes to plot. If NULL, plots all countries.}

\item{save_plots}{Boolean indicator, if set to TRUE plots will be saved in output directory of results fit.}

\item{output_folder}{Folder to save plots in, if save_plots is TRUE, to overwrite where plots are saved.}
}
\description{
Plot model fits, with options to add routine data or additional fits for comparison.
}
