% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subnational_comparison.R
\name{plot_subnational_comparison}
\alias{plot_subnational_comparison}
\title{Plot subnational comparison}
\usage{
plot_subnational_comparison(
  results,
  results2 = NULL,
  model_names = c("Bayesian Model 1", "Bayesian Model 2"),
  year_select,
  ymin_select = 0,
  ymax_select = NA,
  arrange_point = TRUE
)
}
\arguments{
\item{results}{Model fit.}

\item{results2}{Option model fit 2. If NULL, only one model will be plotted.}

\item{model_names}{Optional names for the models to be plotted.
Should be of length 2 if results2 is not NULL, otherwise length 1.}

\item{year_select}{Year to plot. Should be a single value.}

\item{ymin_select}{Minimum y-axis value. Default is 0.}

\item{ymax_select}{Maximum y-axis value. Default is NA (no limit).}

\item{arrange_point}{Boolean to arrange points by median value. Default is TRUE.}
}
\value{
ggplot object with comparison plot
}
\description{
Plot subnational comparison
}
