% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_helpers.R
\name{probit}
\alias{probit}
\title{Probit Transform (Probit to Proportion)}
\usage{
probit(x)
}
\arguments{
\item{x}{Value on probit scale}
}
\value{
Value on proportion scale (between 0 and 1)
}
\description{
Transforms values from probit scale to proportion scale using the standard normal CDF.
}
