% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process survey data for a selected indicator}
\usage{
process_data(dat, regions_dat, indicator = "ideliv", verbose = TRUE)
}
\arguments{
\item{dat}{data frame with iso, year, indic, r (indicator), se, source, final_year columns}

\item{regions_dat}{data frame with iso and cluster columns}

\item{indicator}{character, default "ideliv"}

\item{verbose}{logical, whether to print messages about data processing}
}
\value{
processed data frame
}
\description{
Process survey data for a selected indicator
}
