% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{write model}
\usage{
write_model(add_aggregates = FALSE, add_routine = FALSE)
}
\arguments{
\item{add_aggregates}{Add national aggregates? makes sense only for subnational all-region run for 1 country}

\item{add_routine}{Add routine data?}
}
\value{
writes a stan file to the stan directory, naming ...
}
\description{
write model
}
