% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standata_routine.R
\name{get_standata_routine}
\alias{get_standata_routine}
\title{get_standata_routine}
\usage{
get_standata_routine(
  service_statistic_df,
  hyper_param,
  time_index,
  geo_unit_index
)
}
\arguments{
\item{service_statistic_df}{tibble with  iso; admin1 for subnat; routine; sd_routine; routine_roc; sd_routine_roc; year

filtered to pop (eg country)}

\item{hyper_param}{mean and sd hyperparameters for roc data model}

\item{time_index}{from the model fit}

\item{geo_unit_index}{from the model fit}
}
\value{
list with dat_routine (for plotting) and routine_list (to pass to stan)
}
\description{
get_standata_routine
}
