% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fit.R
\name{load_fit}
\alias{load_fit}
\title{UPDATE: to check if still used.
Load a fit. This fit should be saved in a folder that contains
a file named \code{fit.rds} with the fpemplus object as well as csv files produced
by cmdstanr as a result of the estimation process.}
\usage{
load_fit(path)
}
\arguments{
\item{path}{path to the folder containing the model fit}
}
\value{
object of class fpemplus
}
\description{
UPDATE: to check if still used.
Load a fit. This fit should be saved in a folder that contains
a file named \code{fit.rds} with the fpemplus object as well as csv files produced
by cmdstanr as a result of the estimation process.
}
\keyword{internal}
